refactor(whatsrust): extract message payload builders
